+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/ast/ApexNode.java
@@ -14,8 +14,7 @@ import apex.jorje.semantic.ast.AstNode;
 
 /**
  * Root interface implemented by all Apex nodes. Apex nodes wrap a tree
- * obtained from an external parser (Jorje). The underlying AST node is
- * available with {@link #getNode()}.
+ * obtained from an external parser (Jorje).
  *
  * @param <T> Type of the underlying Jorje node
  */
@@ -31,16 +30,6 @@ public interface ApexNode<T extends AstNode> extends GenericNode<ApexNode<?>> {
     default Object jjtAccept(ApexParserVisitor visitor, Object data) {
         return acceptVisitor(visitor, data);
     }
-
-
-    /**
-     * Get the underlying AST node.
-     * @deprecated the underlying AST node should not be available outside of the AST node.
-     *      If information is needed from the underlying node, then PMD's AST node need to expose
-     *      this information.
-     */
-    @Deprecated
-    T getNode();
 
 
     boolean hasRealLoc();