+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/multifile/ApexMultifileAnalysis.java
@@ -16,7 +16,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import net.sourceforge.pmd.annotation.Experimental;
-import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.apex.ApexLanguageProcessor;
 import net.sourceforge.pmd.lang.apex.ApexLanguageProperties;
 
@@ -52,8 +51,7 @@ public final class ApexMultifileAnalysis {
     }
 
 
-    @InternalApi
-    public ApexMultifileAnalysis(ApexLanguageProperties properties) {
+    ApexMultifileAnalysis(ApexLanguageProperties properties) {
         String rootDir = properties.getProperty(ApexLanguageProperties.MULTIFILE_DIRECTORY);
         LOG.debug("MultiFile Analysis created for {}", rootDir);