+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/multifile/InternalApiBridge.java
@@ -1,0 +1,22 @@
+/*
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+package net.sourceforge.pmd.lang.apex.multifile;
+
+import net.sourceforge.pmd.annotation.InternalApi;
+import net.sourceforge.pmd.lang.apex.ApexLanguageProperties;
+
+/**
+ * @apiNote Internal API
+ */
+@InternalApi
+public final class InternalApiBridge {
+    private InternalApiBridge() {
+        // utility class
+    }
+
+    public static ApexMultifileAnalysis createApexMultiFileAnalysis(ApexLanguageProperties properties) {
+        return new ApexMultifileAnalysis(properties);
+    }
+}