+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/bestpractices/AbstractApexUnitTestRule.java
@@ -2,22 +2,19 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.apex.rule;
+package net.sourceforge.pmd.lang.apex.rule.bestpractices;
 
-import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.apex.ast.ASTModifierNode;
 import net.sourceforge.pmd.lang.apex.ast.ASTUserClass;
 import net.sourceforge.pmd.lang.apex.ast.ApexNode;
+import net.sourceforge.pmd.lang.apex.rule.AbstractApexRule;
 
 /**
  * Do special checks for apex unit test classes and methods
  *
  * @author a.subramanian
- * @deprecated Internal API
  */
-@Deprecated
-@InternalApi
-public abstract class AbstractApexUnitTestRule extends AbstractApexRule {
+abstract class AbstractApexUnitTestRule extends AbstractApexRule {
 
     /**
      * Don't bother visiting this class if it's not a class with @isTest and